Optimize and clean up `newpost()` and initials
I'll be frank, I found the code difficult to read due to how convoluted
it was by its excessive compaction, akin to a wall-of-text. Not nice to
work with and not nice to view.

While having condensed code can be great, I believe you've taken it
too far. There is very little adherence to the traditional formatting
styles of the blocks, so I've sorted that for the initial stuff and for
the `newpost()` function. It's now a lot easier to read and edit.

I've added `FAIL()` (my error message system I use a lot) for cleaner
code and easier additional error messages. The syntax is:

```
FAIL [0|1] [$LINENO] [MESSAGE]
```

Where `0` would be to _not_ exit, and `1` would be to `exit 1`.

Your initial `read` (`ask`) now needs just the one character of input
before continuing (no need to press Enter), and the following case
statement now catches both upper- and lower-case of the 'Y' letter.

I've addressed the repeated use of the logical _and_ operator (`&&`), -
so that things are clearly defined.

You've used `-r` a lot with `read`, yet I see no reason why you would
want the expansion of escape sequences. Unless I've missed something, -
I went ahead and removed the flag.

I've greatly optimized how the `title` is parsed, using just `awk`, -
instead of a multitude of different processes.

I've cleaned up some variables, but that's nothing major.

If there are any bugs, please let me know so that I can fix them, as
with all commits to come.

Awesome idea for a script, by the way!